flask_with_spark/requirements.txt
alembic==0.9.6
amqp==2.3.2
bcrypt==3.1.4
beautifulsoup4==4.6.0
billiard==3.5.0.3
bleach==2.1.3
celery==4.1.0
cffi==1.11.5
chardet==2.3.0
click==6.7
colorama==0.3.2
decorator==4.3.0
dominate==2.3.1
entrypoints==0.2.3
Flask==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-Celery-Helper==1.1.0
Flask-Cors==3.0.3
Flask-Markdown==0.3
Flask-Migrate==2.1.1
flask-mongoengine==0.9.5
Flask-PyMongo==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
freezegun==0.3.10
future==0.16.0
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
kombu==4.1.0
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
mistune==0.8.1
mongoengine==0.15.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.3
numpy==1.14.5
oauth2client==4.1.2
ordereddict==1.1
pandocfilters==1.4.2
parso==0.1.0
py4j==0.10.7
pyasn1==0.4.3
pyasn1-modules==0.2.2
pycparser==2.18
Pygments==2.2.0
pymongo==3.5.1
pyspark==2.3.1
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.5
redis==2.10.6
requests==2.4.3
rsa==3.4.2
six==1.8.0
SQLAlchemy==1.2.9
testpath==0.3.1
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
visitor==0.1.3
webencodings==0.5.1
Werkzeug==0.14.1
WTForms==2.2.1
